Initial implementation of example one-way EdgeX Cloud MQTTS Service
Service is built upon the EdgeX Applications Functions SDK. It receives
northbound events and forwards them to a user-defined topic on an
upstream MQTTS server. When an event references a device this service
hasn't seen before, the service will separately query the EdgeX
core-metadata service for the device's metadata and forward the result
to a separate user-defined topic on an upstream MQTTS server.

Signed-off-by: Michael W. Estrin <me@michaelestrin.com>